Renaming to AcquisitionSeriesProperties because it is a better description.
